LayoutTests/inspector/extensions/extensions-events.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/extensions-test.js"></script>
<script type="text/javascript">

function extension_testOnSelectionChanged(nextTest)
{
    function onSelectionChanged()
    {
        webInspector.panels.elements.onSelectionChanged.removeListener(onSelectionChanged);
        output("onSelectionChanged fired");
        nextTest();
    }
    webInspector.panels.elements.onSelectionChanged.addListener(onSelectionChanged);
    webInspector.inspectedWindow.eval("inspect(document.body.children[0])");
}

function extension_testOnResourceFinished(nextTest)
{
    function onFinished()
    {
        webInspector.resources.onFinished.removeListener(onFinished);
        output("onFinished fired");
        nextTest();
    }
    webInspector.resources.onFinished.addListener(onFinished);
    webInspector.inspectedWindow.eval("var xhr = new XMLHttpRequest(); xhr.open('GET', '" + location.href + "', false); xhr.send(null);");
}

function extension_testPageLoadEvents(nextTest)
{
    var queuedOutput = [];

    // Logging on onLoaded is tricky due to race conditions with output("Page reloaded") in page initialization code.
    // So only log when we're sure all handlers are called and reload is done.
    function queueOutputAndContinue(message)
    {
        queuedOutput.push(message);
        if (queuedOutput.length != 3)
            return;
        for (i = 0; i < queuedOutput.length; ++i) {
            if (queuedOutput[i])
                output(queuedOutput[i]);
        }
        nextTest();
    }

    webInspector.inspectedWindow.onDOMContentLoaded.addListener(function(time) {
        queueOutputAndContinue("onDOMContentLoaded fired: " + typeof time);
    });

    webInspector.inspectedWindow.onLoaded.addListener(function(time) {
        queueOutputAndContinue("onLoaded fired: " + typeof time);
    });
    dispatchOnFrontend({ command: "reload" }, queueOutputAndContinue);
}

function extension_testPanelShowHide(nextTest)
{
    var listenersToCleanup = [];
    function onPanelEvent(type, panelId)
    {
        output("Got " + type + " event for panel " + panelId);
        if (type !== "onShown")
            return;
        if (panelId === "extension") {
            addListener(webInspector.panels.elements, "elements", "onShown");
            showPanel("elements");
        } else if (panelId === "elements") {
            cleanupListeners();
            nextTest();
        }
    }
    function addListener(panel, panelId, type)
    {
        var listener = bind(onPanelEvent, null, type, panelId);
        var event = panel[type];
        listenersToCleanup.push({ event: event, listener: listener });
        event.addListener(listener);
    }
    function cleanupListeners()
    {
        for (var i = 0; i < listenersToCleanup.length; ++i)
            listenersToCleanup[i].event.removeListener(listenersToCleanup[i].listener);
    }
    function onPanelCreated(panel)
    {
        addListener(panel, "extension", "onShown");
        addListener(panel, "extension", "onHidden");
        addListener(webInspector.panels.elements, "elements", "onHidden");
        showPanel("extension");
    }
    showPanel("elements");
    webInspector.panels.create("Test Panel", "extension-panel.png", "extension-panel.html", onPanelCreated);
}

</script>
</head>
<body onload="runTest()">
<p>Tests WebInspector extension API</p>
</body>
</html>
